KernelBench/requirements.txt
anthropic
modal
numpy
openai
packaging
pydra_config
torch==2.5.0
tqdm
datasets
transformers
google-generativeai
together
pytest
ninja
archon-ai
einops